// frontend/src/types/api.ts
export interface User {
  id: number;
  email: string;
  is_active: boolean;
  role: "buyer" | "seller" | "admin";
}

export interface UserCreate {
  email: string;
  password: string;
  role?: "buyer" | "seller";
}

export interface UserAdminCreate {
  email: string;
  password: string;
  role: "buyer" | "seller" | "admin";
}

export interface UserRoleUpdate {
  role: "buyer" | "seller" | "admin";
}

export interface UserFilter {
  search?: string;
  role?: string;
  is_active?: boolean;
  sort_by?: "email" | "role" | "id_asc" | "id_desc";
  page?: number;
  size?: number;
}